+++ jvm/src/test/scala/eval/VariableSpec.scala
@@ -41,14 +41,14 @@ class VariableSpec extends FunSpec with Matchers {
   }
 
   it("should produce an accurate variable map with buffer in a simple case") {
-    Vars.varsWithBuffer(FocalMax(List(RasterVar("someRaster")), Square(1))) should be (Map("someRaster" -> (MamlKind.Image, 1)))
+    Vars.varsWithBuffer(FocalMax(List(RasterVar("someRaster")), Square(1), TargetCell.All)) should be (Map("someRaster" -> (MamlKind.Image, 1)))
   }
 
   it("should produce an accurate variable map with buffer in an ambiguous case") {
     val ast = Addition(List(
         FocalMax(List(
-          FocalMax(List(RasterVar("someRaster")), Square(1))
-        ), Square(1)),
+          FocalMax(List(RasterVar("someRaster")), Square(1), TargetCell.All)
+        ), Square(1), TargetCell.All),
         RasterVar("someRaster")
     ))